chore: 교환 게시물 삭제 시, 권한 확인 추가

+++ src/main/java/com/CSP2/switchcon/exchange/service/ExchangePostService.java
@@ -103,6 +103,9 @@ public class ExchangePostService {
         ExchangePost exchangePost = exchangePostRepository.findById(exchangePostId)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorCode.EXCHANGE_POST_NOT_FOUND));
 
+        if (!exchangePost.getGifticon().getMember().getId().equals(member.getId()))
+            throw new BusinessException(ErrorCode.FORBIDDEN_DELETE_EXCHANGE_POST);
+
         exchangePostRepository.deleteByIdAndMember(member, exchangePostId);
 
         exchangePost.getGifticon().updateActive(true);
